Remove redundant tests in java/security
